moved more functions to parser test
